Move data directory function to utils package inside adhoc.
This way both the client and the server will be able to use it.
