+++ pkg/adhoc/util/util.go
@@ -1,0 +1,7 @@
+package util
+
+import "path/filepath"
+
+func DataDirectory() string {
+	return filepath.Join(dataBaseDirectory(), "pyroscope")
+}
