+++ pkg/adhoc/util/util_darwin.go
@@ -1,4 +1,4 @@
-package adhoc
+package util
 
 import (
 	"os"
@@ -6,12 +6,9 @@ import (
 )
 
 func dataBaseDirectory() string {
-	if dir, ok := os.LookupEnv("XDG_DATA_HOME"); ok {
-		return dir
-	}
 	homeDir, ok := os.LookupEnv("HOME")
 	if !ok {
 		homeDir = "/"
 	}
-	return filepath.Join(homeDir, ".local", "share")
+	return filepath.Join(homeDir, "Library", "Application Support")
 }
